Setup accuracy benchmarks for all.
